feat: add support for converting flattened nested trial hp to experiment config hp
