+++ webui/react/src/utils/experiment.ts
@@ -1,6 +1,7 @@
 import {
   ExperimentBase, ExperimentSearcherName, Hyperparameter,
-  Hyperparameters, HyperparametersFlattened,
+  HyperparameterBase,
+  Hyperparameters, HyperparametersFlattened, HyperparameterType, TrialHyperparameters,
 } from 'types';
 
 export const flattenHyperparameters = (
@@ -22,4 +23,27 @@ export const flattenHyperparameters = (
 export const isSingleTrialExperiment = (experiment: ExperimentBase): boolean => {
   return experiment?.config.searcher.name === ExperimentSearcherName.Single
         || experiment?.config.searcher.max_trials === 1;
+};
+
+export const trialHParamsToExperimentHParams = (
+  trialHParams: TrialHyperparameters,
+): Hyperparameters => {
+  const experimentHParams: Hyperparameters = {};
+  Object.entries(trialHParams).forEach(([ paramPath, value ]) => {
+    let key = paramPath;
+    let matches = key.match(/^([^.]+)\.(.+)$/);
+    let pathRef: Hyperparameters | HyperparameterBase = experimentHParams;
+    while (matches?.length === 3) {
+      const prefix = matches[1];
+      key = matches[2];
+      (pathRef as Hyperparameters)[prefix] = (pathRef as Hyperparameters)[prefix] ?? {};
+      pathRef = (pathRef as Hyperparameters)[prefix];
+      matches = key.match(/^([^.]+)\.(.+)$/);
+    }
+    (pathRef as Hyperparameters)[key] = {
+      type: HyperparameterType.Constant,
+      val: value as number,
+    };
+  });
+  return experimentHParams;
 };
